% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DOE_data.r
\name{get_DOE_data}
\alias{get_DOE_data}
\title{get_DOE_data}
\usage{
get_DOE_data(
  begin_date,
  end_date,
  flow_site,
  flow_params,
  daily_or_instantaneous
)
}
\arguments{
\item{begin_date}{.}

\item{end_date}{.}

\item{flow_site}{.}

\item{flow_params}{.}

\item{daily_or_instantaneous}{.}
}
\value{
.
}
\description{
Downloads data from the Washington State Department of Ecology's Freshwater DataStream
}
