% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_USGS_data.r
\name{get_usgs_data}
\alias{get_usgs_data}
\title{get_usgs_data}
\usage{
get_usgs_data(
  flow_site,
  flow_params,
  daily_or_instantaneous,
  begin_date,
  end_date
)
}
\arguments{
\item{flow_site}{.}

\item{flow_params}{.}

\item{daily_or_instantaneous}{.}

\item{begin_date}{.}

\item{end_date}{.}
}
\value{
.
}
\description{
Downloads Streamflow Data from the USGS National Water Information System (NWIS)
}
