% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_or_load_pack.r
\name{install_or_load_pack}
\alias{install_or_load_pack}
\title{install_or_load_pack}
\usage{
install_or_load_pack(pack)
}
\arguments{
\item{pack}{.}
}
\value{
.
}
\description{
Install and Load Packages
}
